pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h2pcontrol"
version = "1.0"
description = "Project initialized by h2pcontrol cli"
authors = [
  { name = "Tijmen Hilgenkamp", email = "t.j.hilgenkamp@student.vu.nl" },
]
requires-python = ">=3.11"
dependencies = [
    "betterproto2==0.4.0",
    "grpcio>=1.72.1",
    "grpclib>=0.4.8",
]


# [tool.setuptools]
# packages = ["h2pcontrol", "h2pcontrol.pb"]
